requirements.txt
casadi==3.6.3
catkin_pkg==0.5.2
dynamic_reconfigure==1.7.1
ipdb==0.13.13
matplotlib==3.1.2
numpy==1.24.4
numpy==1.17.4
pandas==2.0.3
psutil==5.9.4
PyAutoGUI==0.9.54
PyYAML==6.0.1
rosgraph_msgs==0.0.1
roslaunch==1.15.9
rosparam==1.15.9
rospy==1.15.9
sensor_msgs==1.13.0
statsmodels==0.14.0
std_msgs==0.0.1
sympy==1.12
tf==1.13.2
tqdm==4.66.1
